sdlc: override default CADDYPATH to activate telemetry
The default caddypath is `~/.caddy`, and this path is
on a read-only filesystem in this test harness. This
means telemetry silently fails to activate.

After this commit, use `/var/www/` as the caddypath
since this is a writable filesystem in the test harness.
The caddy user also has write permissions on this path.