ci: simplify how to build fixtures
Before this commit, we had to create a fixtures container
outside of docker-compose. This complicated the steps to
build and clean, plus it prevented docker-compose from
doing a clean restart of caddy1 when the fixtures changed.

After this commit, do everything within docker-compose
so that docker-compose can track state of caddy1.